create conda/common/disk.py
