+++ conda/fetch.py
@@ -29,7 +29,8 @@ from .exceptions import (ProxyError, CondaRuntimeError, CondaSignatureError, Con
                          MD5MismatchError)
 from .install import add_cached_package, find_new_location, package_cache, dist2pair, rm_rf
 from .lock import FileLock
-from .utils import exp_backoff_fn, memoized
+from .utils import memoized
+from conda.common.disk import exp_backoff_fn, rm_rf
 
 log = getLogger(__name__)
 dotlog = getLogger('dotupdate')
