+++ conda/plan.py
@@ -24,7 +24,8 @@ from .exceptions import (InstallError, RemoveError, CondaIndexError,
 from .history import History
 from .install import (dist2quad, LINK_HARD, link_name_map, name_dist, is_fetched,
                       is_extracted, is_linked, find_new_location, dist2filename, LINK_COPY,
-                      LINK_SOFT, try_hard_link, rm_rf)
+                      LINK_SOFT, try_hard_link)
+from conda.common.disk import rm_rf
 from .resolve import MatchSpec, Resolve, Package
 from .utils import md5_file, human_bytes, on_win
