+++ conda/utils.py
@@ -1,19 +1,12 @@
 from __future__ import absolute_import, division, print_function
 
 import collections
-import errno
 import hashlib
 import logging
 import re
-from shutil import rmtree
 import sys
 import threading
-import time
 from functools import partial
-from itertools import chain
-from os import W_OK, access, chmod, getpid, listdir, unlink, walk, stat
-from os.path import basename, exists, isdir, join, dirname
-from stat import S_IWRITE, S_IEXEC
 
 from .common.url import path_to_url
 
@@ -93,106 +86,6 @@ def gnu_get_libc_version():
     f = libc.gnu_get_libc_version
     f.restype = c_char_p
     return f()
-
-
-def try_write(dir_path, heavy=False):
-    """Test write access to a directory.
-
-    Args:
-        dir_path (str): directory to test write access
-        heavy (bool): Actually create and delete a file, or do a faster os.access test.
-           https://docs.python.org/dev/library/os.html?highlight=xattr#os.access
-
-    Returns:
-        bool
-
-    """
-    if not isdir(dir_path):
-        return False
-    if on_win or heavy:
-        # try to create a file to see if `dir_path` is writable, see #2151
-        temp_filename = join(dir_path, '.conda-try-write-%d' % getpid())
-        try:
-            with open(temp_filename, mode='wb') as fo:
-                fo.write(b'This is a test file.\n')
-            backoff_unlink(temp_filename)
-            return True
-        except (IOError, OSError):
-            return False
-        finally:
-            backoff_unlink(temp_filename)
-    else:
-        return access(dir_path, W_OK)
-
-
-def backoff_unlink(file_or_symlink_path):
-    def _unlink(path):
-        make_writable(path)
-        unlink(path)
-
-    try:
-        exp_backoff_fn(lambda f: exists(f) and _unlink(f), file_or_symlink_path)
-    except (IOError, OSError) as e:
-        if e.errno not in (errno.ENOENT,):
-            # errno.ENOENT File not found error / No such file or directory
-            raise
-
-
-def backoff_rmdir(dirpath):
-    if not isdir(dirpath):
-        return
-
-    # shutil.rmtree:
-    #   if onerror is set, it is called to handle the error with arguments (func, path, exc_info)
-    #     where func is os.listdir, os.remove, or os.rmdir;
-    #     path is the argument to that function that caused it to fail; and
-    #     exc_info is a tuple returned by sys.exc_info() ==> (type, value, traceback).
-    def retry(func, path, exc_info):
-        if getattr(exc_info[1], 'errno', None) == errno.ENOENT:
-            return
-        recursive_make_writable(dirname(path))
-        func(path)
-
-    def rmdir(path):
-        try:
-            recursive_make_writable(path)
-            exp_backoff_fn(rmtree, path, onerror=retry)
-        except (IOError, OSError) as e:
-            if e.errno == errno.ENOENT:
-                log.debug("no such file or directory: %s", path)
-            else:
-                raise
-
-    for root, dirs, files in walk(dirpath, topdown=False):
-        for file in files:
-            backoff_unlink(join(root, file))
-        for dir in dirs:
-            rmdir(join(root, dir))
-
-    rmdir(dirpath)
-
-
-def make_writable(path):
-    st = stat(path)
-    chmod(path, st.st_mode | S_IWRITE | S_IEXEC if isdir(path) else st.st_mode | S_IWRITE)
-
-
-def recursive_make_writable(path):
-    # The need for this function was pointed out at
-    #   https://github.com/conda/conda/issues/3266#issuecomment-239241915
-    # Especially on windows, file removal will often fail because it is marked read-only
-    if isdir(path):
-        for root, dirs, files in walk(path):
-            for path in chain.from_iterable((files, dirs)):
-                try:
-                    exp_backoff_fn(make_writable, join(root, path))
-                except (IOError, OSError) as e:
-                    if e.errno == errno.ENOENT:
-                        log.debug("no such file or directory: %s", path)
-                    else:
-                        raise
-    else:
-        exp_backoff_fn(make_writable, path)
 
 
 def hashsum_file(path, mode='md5'):
@@ -398,41 +291,6 @@ else:
             pathsep=" ",
                     ),
     }
-
-
-def exp_backoff_fn(fn, *args, **kwargs):
-    """Mostly for retrying file operations that fail on Windows due to virus scanners"""
-    if not on_win:
-        return fn(*args, **kwargs)
-
-    import random
-    # with max_tries = 6, max total time ~= 3.2 sec
-    # with max_tries = 7, max total time ~= 6.5 sec
-    max_tries = 7
-    for n in range(max_tries):
-        try:
-            result = fn(*args, **kwargs)
-        except (OSError, IOError) as e:
-            log.debug(repr(e))
-            if e.errno in (errno.EPERM, errno.EACCES):
-                if n == max_tries-1:
-                    raise
-                sleep_time = ((2 ** n) + random.random()) * 0.1
-                caller_frame = sys._getframe(1)
-                log.debug("retrying %s/%s %s() in %g sec",
-                          basename(caller_frame.f_code.co_filename),
-                          caller_frame.f_lineno, fn.__name__,
-                          sleep_time)
-                time.sleep(sleep_time)
-            elif e.errno in (errno.ENOENT,):
-                # errno.ENOENT File not found error / No such file or directory
-                raise
-            else:
-                log.error("Uncaught backoff with errno %d", e.errno)
-                raise
-        else:
-            return result
-
 
 # put back because of conda build
 urlpath = url_path = path_to_url
